+++ src/containers/SettingsMenu/index.js
@@ -1,1 +1,2 @@
+/* eslint-disable import/prefer-default-export */
 export { default as SettingsMenu } from './SettingsMenu';
