+++ src/ducks/modules/ui.js
@@ -3,7 +3,6 @@
  */
 
 const initialState = {
-  isMenuOpen: false,
   protocolIndex: 0,
 };
