+++ src/styles/components/timeline/_timeline-stage.scss
@@ -21,9 +21,9 @@
   }
 
   &__label {
-    flex: 0 0 auto;
     display: flex;
     align-items: center;
+    font-weight: 700;
   }
 }
