+++ src/styles/containers/_alter-form-interface.scss
@@ -51,7 +51,6 @@
         overflow-x: hidden;
         overflow-y: auto;
         flex-direction: column;
-        max-width: 55rem; // same as ego form max width
 
         h1 {
           text-align: center;
